Almost finished. Needs end story, attack fixes, enemy ai fixes, end screen, highscore file

+++ src/javaapplication5/Wall.java
@@ -11,43 +11,47 @@ import processing.core.PImage;
  */
 public class Wall {
     private PApplet app;
-    private float x, y;
+    private int x, y;
     private int width, height;
     private PImage wall;
     
-    public Wall(PApplet app, float x, float y, int width, int height, String imagePath) {
+    // Constructor
+    public Wall(PApplet app, int x, int y, int width, int height, String imagePath) {
         this.app = app;
         this.x = x;
         this.y = y;
         this.width = width;
         this.height = height;
         
+        // Tries to load wall image
         try {
             this.wall = app.loadImage(imagePath);
         } catch (Exception e) {
             this.wall = null;
         }
     }
-    
+    // Draws wall
     public void draw() {
         if (wall != null) {
             app.image(wall, x, y, width, height);
         } else {
+            // colour if image fails to load
             app.fill(210, 180, 140);
             app.rect(x, y, width, height);
         }
     }
     
-    public boolean isCollidingWith(float objX, float objY, int objWidth, int objHeight) {
+    // Checks for collision
+    public boolean isCollidingWith(int objX, int objY, int objWidth, int objHeight) {
         return objX < x + width && objX + objWidth > x && objY < y + height && objY + objHeight > y;
     }
     
-    // Getter
-    public float getX() {
+    // Getters
+    public int getX() {
         return x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
